# docs/swagger.yaml
basePath: /
definitions:
...
  utils.BadRequest:
    properties:
      error:
        example: string
        type: string
      message:
        example: message
        type: string
      result:
        example: "null"
        type: string
      status:
        example: 400
        type: integer
      success:
        example: false
        type: boolean
    type: object
...
  utils.InternalServerError:
    properties:
      error:
        example: message
        type: string
      message:
        example: message
        type: string
      result:
        example: "null"
        type: string
      status:
        example: 500
        type: integer
      success:
        example: false
        type: boolean
    type: object
  utils.Invoice:
    properties:
      attach:
        example: path/to/file.pdf or file
        type: string
      client_name:
        example: John Doe
        type: string
      to:
        example: mail@example.com
        type: string
      url_upload:
        example: https://example.com/upload
        type: string
    type: object
...
  utils.Registered:
    properties:
      client_name:
        example: John Doe
        type: string
      to:
        example: mail@example.com
        type: string
    type: object
  utils.Reject:
    properties:
      client_name:
        example: John Doe
        type: string
      reason:
        example: this is my reason
        type: string
      to:
        example: mail@example.com
        type: string
    type: object
  utils.Success:
    properties:
      error:
        example: "null"
        type: string
      message:
        example: message
        type: string
      result:
        example: message
        type: string
      status:
        example: 200
        type: integer
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: Documentation for Postmail
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Postmail API
  version: "1.0"
paths:
...
  /send/template/invoice:
    post:
      consumes:
      - application/json
      description: Using template Invoice Paid
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.Invoice'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.BadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.InternalServerError'
      security:
      - X-Auth-Key: []
      summary: Send email template - Invoice Paid
      tags:
      - Send Template
...
  /send/template/registered:
    post:
      consumes:
      - application/json
      description: Using template Registered
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.Registered'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.BadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.InternalServerError'
      security:
      - X-Auth-Key: []
      summary: Send email template - Registered
      tags:
      - Send Template
  /send/template/rejected:
    post:
      consumes:
      - application/json
      description: Using template Rejected
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.Reject'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.BadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.InternalServerError'
      security:
      - X-Auth-Key: []
      summary: Send email template - Rejected
      tags:
      - Send Template
...
securityDefinitions:
  X-Auth-Key:
    description: Input your token authorized
    in: header
    name: X-Auth-Key
    type: apiKey
swagger: "2.0"